load lens app lazily (#64769)
